+++ src/main/java/com/assistance/DogShelter/controller/VolunteerController.java
@@ -42,7 +42,7 @@ public class VolunteerController {
     @Operation(summary = "Add volunteer",
             description = "Adds a new volunteer to the system",
             responses = {
-                    @ApiResponse(responseCode = "200", description = "Successfully created"),
+                    @ApiResponse(responseCode = "201", description = "Successfully created"),
                     @ApiResponse(responseCode = "400", description = "Invalid request")
             })
     public Volunteer addVolunteer(@RequestBody Volunteer volunteer) {
@@ -60,17 +60,14 @@ public class VolunteerController {
     @Operation(summary = "Search volunteer",
             description = "Searches for a volunteer in the system",
             responses = {
-                    @ApiResponse(responseCode = "200", description = "Success"),
+                    @ApiResponse(responseCode = "200", description = "Successfully removed"),
                     @ApiResponse(responseCode = "400", description = "Invalid request"),
                     @ApiResponse(responseCode = "404", description = "Volunteer not found")
             })
-    public Volunteer findVolunteerById(@PathVariable Long id) {
+    public ResponseEntity<Volunteer> findVolunteerById(@PathVariable Long id) {
         Optional<Volunteer> volunteer = volunteerService.findVolunteerById(id);
-
-        if (volunteer.isEmpty()) {
-            return (Volunteer) ResponseEntity.status(HttpStatus.NOT_FOUND).build().getBody();
-        }
-        return ResponseEntity.status(HttpStatus.FOUND).body(findVolunteerById(id)).getBody();
+        return volunteer.map(value -> ResponseEntity.status(HttpStatus.OK).body(value))
+                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
 
     /**
@@ -84,16 +81,17 @@ public class VolunteerController {
     @Operation(summary = "Update volunteer",
             description = "Updates a volunteer's information",
             responses = {
-                    @ApiResponse(responseCode = "200", description = "Successfully updated"),
-                    @ApiResponse(responseCode = "400", description = "Invalid request")
+                    @ApiResponse(responseCode = "200", description = "Successfully removed"),
+                    @ApiResponse(responseCode = "400", description = "Invalid request"),
+                    @ApiResponse(responseCode = "404", description = "Volunteer not found")
             })
-    public Volunteer editVolunteer(@RequestBody Volunteer volunteer, @PathVariable Long id) {
-        Volunteer foundVolunteer = volunteerService.editVolunteer(volunteer);
-
-        if (foundVolunteer == null) {
-            return (Volunteer) ResponseEntity.status(HttpStatus.NOT_FOUND).build().getBody();
+    public ResponseEntity<Volunteer> editVolunteer(@RequestBody Volunteer volunteer, @PathVariable Long id) {
+        Optional<Volunteer> foundVolunteer = volunteerService.findVolunteerById(id);
+        if (foundVolunteer.isPresent()) {
+            Volunteer updatedVolunteer = volunteerService.editVolunteer(volunteer);
+            return ResponseEntity.status(HttpStatus.OK).body(updatedVolunteer);
         }
-        return ResponseEntity.status(HttpStatus.FOUND).body(findVolunteerById(id)).getBody();
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     /**
@@ -107,12 +105,17 @@ public class VolunteerController {
     @Operation(summary = "Set volunteer",
             description = "Set status for a volunteer as busy",
             responses = {
-                    @ApiResponse(responseCode = "200", description = "Successfully removed"),
-                    @ApiResponse(responseCode = "400", description = "Invalid request")
+                    @ApiResponse(responseCode = "200", description = "Successfully updated"),
+                    @ApiResponse(responseCode = "400", description = "Invalid request"),
+                    @ApiResponse(responseCode = "404", description = "Volunteer not found")
             })
-    public Volunteer setVolunteerBusy(@PathVariable Long userId, Boolean busy) {
-        Volunteer setVolunteer = volunteerService.keepVolunteerBusy(userId, busy);
-        return ResponseEntity.status(HttpStatus.FOUND).body(setVolunteer).getBody();
+    public ResponseEntity<Volunteer> setVolunteerBusy(@PathVariable Long userId, @RequestParam Boolean busy) {
+        Optional<Volunteer> volunteer = volunteerService.findVolunteerById(userId);
+        if (volunteer.isPresent()) {
+            Volunteer updatedVolunteer = volunteerService.keepVolunteerBusy(userId, busy);
+            return ResponseEntity.status(HttpStatus.OK).body(updatedVolunteer);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     /**
@@ -125,12 +128,16 @@ public class VolunteerController {
     @Operation(summary = "Remove volunteer",
             description = "Removes a volunteer from the system",
             responses = {
-                    @ApiResponse(responseCode = "200", description = "Successfully removed"),
-                    @ApiResponse(responseCode = "400", description = "Invalid request")
+                    @ApiResponse(responseCode = "200", description = "Successfully updated"),
+                    @ApiResponse(responseCode = "400", description = "Invalid request"),
+                    @ApiResponse(responseCode = "404", description = "Volunteer not found")
             })
-    public Volunteer deleteVolunteer(@PathVariable Long id) {
-        volunteerService.deleteVolunteer(id);
-        return (Volunteer) ResponseEntity.status(HttpStatus.OK);
+    public ResponseEntity<Void> deleteVolunteer(@PathVariable Long id) {
+        if (volunteerService.findVolunteerById(id).isPresent()) {
+            volunteerService.deleteVolunteer(id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
 //    /**
